requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
apturl==0.5.2
asgiref==3.5.0
async-timeout==4.0.1
attrs==21.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.5.0
duplicity==0.8.21
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.0
fasteners==0.14.1
filelock==3.13.1
Flask==1.1.2
Flask-Cors==3.0.8
frozenlist==1.2.0
future==0.18.2
greenlet==3.0.3
h11==0.13.0
httpcore==1.0.2
httplib2==0.20.2
httptools==0.2.0
httpx==0.26.0
idna==2.6
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==2.11.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==8.10.0
multidict==5.1.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
orjson==3.9.12
package-name==0.1
packaging==21.3
PAM==0.4.2
paramiko==2.9.3
passlib==1.7.4
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.0
protobuf==3.12.4
proton-core==0.1.16
proton-keyring-linux==0.0.1
proton-keyring-linux-secretservice==0.0.1
proton-vpn-api-core==0.20.3
proton-vpn-connection==0.11.3
proton-vpn-gtk-app==4.1.10
proton-vpn-killswitch==0.2.0
proton-vpn-killswitch-network-manager==0.2.0
proton-vpn-logger==0.2.1
proton-vpn-network-manager==0.3.3
proton-vpn-network-manager-openvpn==0.0.4
proton-vpn-session==0.6.5
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.1
pycairo==1.20.1
pycodestyle==2.6.0
pycparser==2.21
pycups==2.0.1
pydantic==2.6.1
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.18.4
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
systemd-python==234
tabulate==0.8.9
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.9.0
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.7
uvicorn==0.15.0
uvloop==0.19.0
virtualenv==20.25.0
wadllib==1.3.6
watchgod==0.8.2
websockets==12.0
Werkzeug==3.0.1
wsproto==1.0.0
xdg==5
xkit==0.0.0
yarl==1.7.2
zipp==1.0.0
